requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
attrs==24.2.0
black==24.10.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
yarl==1.17.1
